fix: use .half().to(device) for all frozen models — converts params AND buffers uniformly

+++ training/train_controlnet.py
@@ -210,10 +210,10 @@ def train(config: dict) -> None:
     controlnet, optimizer, loader, lr_scheduler = accelerator.prepare(
         controlnet, optimizer, loader, lr_scheduler
     )
-    vae = vae.to(accelerator.device, dtype=torch.float16)
-    unet = unet.to(accelerator.device)
-    text_encoder = text_encoder.to(accelerator.device)
-    text_encoder_2 = text_encoder_2.to(accelerator.device)
+    vae = vae.half().to(accelerator.device)
+    unet = unet.half().to(accelerator.device)
+    text_encoder = text_encoder.half().to(accelerator.device)
+    text_encoder_2 = text_encoder_2.half().to(accelerator.device)
 
     global_step = 0
     max_steps = config["training"]["max_train_steps"]
